# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse:
    properties:
      details:
        description: |-
          Детали ошибки (опционально)
          example: Login field is required
        type: string
      error:
        description: |-
          Текст ошибки
          example: Invalid request
        type: string
    type: object
...
  github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.TypeRequest:
    properties:
      name:
        description: |-
          Название типа
          required: true
          example: Йога
        type: string
    type: object
  github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.TypeResponse:
    properties:
      created_at:
        description: |-
          Дата создания
          example: 02.01.2006
        type: string
      id:
        description: |-
          ID типа
          example: 1
        type: number
      name:
        description: |-
          Название типа
          example: Йога
        type: string
    type: object
...
host: https://api.7375.org
info:
  contact:
    email: info@7375.org
    name: Sergei
  description: API для управления видео-контентом BodyBalance.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: BodyBalance API
  version: "1.0"
paths:
...
  /admin/type:
    get:
      description: Возвращает список всех типов контента в системе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.TypeResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Получить все типы
      tags:
      - Admin Types
    post:
      consumes:
      - application/json
      description: Добавляет новый тип контента в систему
      parameters:
      - description: Данные типа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.TypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Тип успешно создан
          schema:
            properties:
              id:
                type: integer
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Создать новый тип
      tags:
      - Admin Types
  /admin/type/{id}:
    delete:
      description: Удаляет тип контента из системы
      parameters:
      - description: ID типа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Тип успешно удален
          schema:
            properties:
              id:
                type: integer
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Удалить тип
      tags:
      - Admin Types
    get:
      description: Возвращает информацию о типе контента по его идентификатору
      parameters:
      - description: ID типа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.TypeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Получить тип по ID
      tags:
      - Admin Types
    put:
      consumes:
      - application/json
      description: Обновляет информацию о существующем типе контента
      parameters:
      - description: ID типа
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные типа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.TypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Тип успешно обновлен
          schema:
            properties:
              id:
                type: integer
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_langowen_bodybalance-backend_internal_http-server_admin_admResponse.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Обновить тип
      tags:
      - Admin Types
...
swagger: "2.0"